pyproject.toml
[tool.poetry]
name = "test-task-DTEAM"
version = "0.1.0"
description = "A Django-based CV management system"
authors = ["Ivan Zashchyk <ivan.zashchyk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11.5"
django = "^5.2.1"
psycopg2-binary = "^2.9.10"
pytest-django = "^4.11.1"
xhtml2pdf = "^0.2.15"
djangorestframework = "^3.16.0"
python-dotenv = "^1.1.0"
gunicorn = "^23.0.0"
celery = {extras = ["redis"], version = "^5.5.2"}
asgiref = "^3.7.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"